Add error handling for DELETE /api/comments/:comment_id endpoint

+++ models/comments.models.js
@@ -52,11 +52,22 @@ exports.createComment = (article_id, body, username, reqBodyLength) => {
 };
 
 exports.removeComment = (comment_id) => {
+  if (isNaN(Number.parseInt(comment_id))) {
+    return Promise.reject({ status: 400, msg: "Bad request" });
+  }
   return db
-    .query("DELETE FROM comments WHERE comment_id = $1 RETURNING *", [
-      comment_id,
-    ])
-    .then((res) => {
-      return res.rows;
+    .query("SELECT * FROM comments WHERE comment_id = $1", [comment_id])
+    .then((z) => {
+      if (z.rows.length === 0) {
+        return Promise.reject({ status: 404, msg: "Not found" });
+      } else {
+        return db
+          .query("DELETE FROM comments WHERE comment_id = $1 RETURNING *", [
+            comment_id,
+          ])
+          .then((res) => {
+            return res.rows;
+          });
+      }
     });
 };
